refactor scenario
